+++ lib/glimmer/dsl/libui/control_expression.rb
@@ -48,4 +48,4 @@ end
 
 # TODO Consider moving all controls underneath Control namespace
 require 'glimmer/libui/control_proxy'
-Dir[File.expand_path('../../libui/*.rb', __dir__)].each {|f| require f}
+Dir[File.expand_path('../../libui/control_proxy/*.rb', __dir__)].each {|f| require f}