+++ lib/glimmer/libui/control_proxy/font_button_proxy.rb
@@ -19,38 +19,52 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/dsl/expression'
-require 'glimmer/dsl/parent_expression'
+require 'glimmer/libui/control_proxy'
 
 module Glimmer
-  module DSL
-    module Libui
-      class ShapeExpression < Expression
-        include ParentExpression
-  
-        def can_interpret?(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.exists?(keyword) and
-            (
-              parent.is_a?(Glimmer::LibUI::ControlProxy::PathProxy) or
-                parent.is_a?(Glimmer::LibUI::Shape)
-            )
-        end
-  
-        def interpret(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.create(keyword, parent, args, &block)
+  module LibUI
+    class ControlProxy
+      # Proxy for LibUI font button objects
+      #
+      # Follows the Proxy Design Pattern
+      class FontButtonProxy < ControlProxy
+        def font
+          @font_descriptor ||= ::LibUI::FFI::FontDescriptor.malloc
+          ::LibUI.font_button_font(@libui, @font_descriptor)
+          {
+            family: @font_descriptor.Family.to_s,
+            size: @font_descriptor.Size,
+            weight: @font_descriptor.Weight,
+            italic: @font_descriptor.Italic,
+            stretch: @font_descriptor.Stretch
+          }
         end
         
-        def add_content(parent, keyword, *args, &block)
+        def family
+          font[:family]
+        end
+        
+        def size
+          font[:size]
+        end
+        
+        def weight
+          font[:weight]
+        end
+        
+        def italic
+          font[:italic]
+        end
+        
+        def stretch
+          font[:stretch]
+        end
+        
+        def destroy
+          ::LibUI.free_font_button_font(@font_descriptor) unless @font_descriptor.nil?
           super
-          parent.post_add_content
         end
-        
       end
     end
   end
 end
-
-# TODO Consider moving all shapes underneath Shape namespace
-require 'glimmer/libui/control_proxy/path_proxy'
-require 'glimmer/libui/shape'
-Dir[File.expand_path('../../libui/shape/*.rb', __dir__)].each {|f| require f}