+++ lib/glimmer/libui/control_proxy/image_part_proxy.rb
@@ -19,38 +19,31 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/dsl/expression'
-require 'glimmer/dsl/parent_expression'
+require 'glimmer/libui/control_proxy'
 
 module Glimmer
-  module DSL
-    module Libui
-      class ShapeExpression < Expression
-        include ParentExpression
-  
-        def can_interpret?(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.exists?(keyword) and
-            (
-              parent.is_a?(Glimmer::LibUI::ControlProxy::PathProxy) or
-                parent.is_a?(Glimmer::LibUI::Shape)
-            )
-        end
-  
-        def interpret(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.create(keyword, parent, args, &block)
-        end
-        
-        def add_content(parent, keyword, *args, &block)
-          super
-          parent.post_add_content
+  module LibUI
+    class ControlProxy
+      # Proxy for LibUI image part objects
+      #
+      # Follows the Proxy Design Pattern
+      class ImagePartProxy < ControlProxy
+        def initialize(keyword, parent, args, &block)
+          @keyword = keyword
+          @parent_proxy = parent
+          @args = args
+          @block = block
+          @enabled = true
+          @children = []
+          post_add_content if @block.nil?
         end
         
+        private
+        
+        def build_control
+          @libui = @parent_proxy.append(*@args)
+        end
       end
     end
   end
 end
-
-# TODO Consider moving all shapes underneath Shape namespace
-require 'glimmer/libui/control_proxy/path_proxy'
-require 'glimmer/libui/shape'
-Dir[File.expand_path('../../libui/shape/*.rb', __dir__)].each {|f| require f}