+++ lib/glimmer/libui/control_proxy/image_proxy.rb
@@ -19,38 +19,47 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/dsl/expression'
-require 'glimmer/dsl/parent_expression'
+require 'glimmer/libui/control_proxy'
+require 'glimmer/data_binding/observer'
+
+using ArrayIncludeMethods
 
 module Glimmer
-  module DSL
-    module Libui
-      class ShapeExpression < Expression
-        include ParentExpression
-  
-        def can_interpret?(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.exists?(keyword) and
-            (
-              parent.is_a?(Glimmer::LibUI::ControlProxy::PathProxy) or
-                parent.is_a?(Glimmer::LibUI::Shape)
-            )
-        end
-  
-        def interpret(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.create(keyword, parent, args, &block)
+  module LibUI
+    class ControlProxy
+      # Proxy for LibUI image objects
+      #
+      # Follows the Proxy Design Pattern
+      class ImageProxy < ControlProxy
+        def initialize(keyword, parent, args, &block)
+          @keyword = keyword
+          @parent_proxy = parent
+          @args = args
+          @block = block
+          @enabled = true
+          @children = []
+          post_add_content if @block.nil?
         end
         
-        def add_content(parent, keyword, *args, &block)
+        def post_add_content
+          build_control
           super
-          parent.post_add_content
+        end
+   
+        def post_initialize_child(child)
+          @children << child
         end
         
+        private
+        
+        def build_control
+          @args = [@children.first.args[1], @children.first.args[2]] if @children.size == 1 && (@args[0].nil? || @args[1].nil?)
+          super
+          @libui.tap do
+            @children.each {|child| child&.send(:build_control) }
+          end
+        end
       end
     end
   end
 end
-
-# TODO Consider moving all shapes underneath Shape namespace
-require 'glimmer/libui/control_proxy/path_proxy'
-require 'glimmer/libui/shape'
-Dir[File.expand_path('../../libui/shape/*.rb', __dir__)].each {|f| require f}