+++ lib/glimmer/libui/control_proxy/quit_menu_item_proxy.rb
@@ -19,38 +19,46 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/dsl/expression'
-require 'glimmer/dsl/parent_expression'
+require 'glimmer/libui/control_proxy/menu_item_proxy'
 
 module Glimmer
-  module DSL
-    module Libui
-      class ShapeExpression < Expression
-        include ParentExpression
-  
-        def can_interpret?(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.exists?(keyword) and
-            (
-              parent.is_a?(Glimmer::LibUI::ControlProxy::PathProxy) or
-                parent.is_a?(Glimmer::LibUI::Shape)
-            )
+  module LibUI
+    class ControlProxy
+      # Proxy for LibUI quit menu item object
+      #
+      # Follows the Proxy Design Pattern
+      class QuitMenuItemProxy < MenuItemProxy
+        def can_handle_listener?(listener_name)
+          listener_name == 'on_clicked' || super
         end
-  
-        def interpret(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.create(keyword, parent, args, &block)
+      
+        def handle_listener(listener_name, &listener)
+          if listener_name == 'on_clicked'
+            @default_behavior_listener = Proc.new do
+              return_value = listener.call(self)
+              if return_value.is_a?(Numeric)
+                return_value
+              else
+                destroy
+                ::LibUI.quit
+                0
+              end
+            end
+            ::LibUI.on_should_quit(&@default_behavior_listener)
+          end
         end
+      
+        private
         
-        def add_content(parent, keyword, *args, &block)
-          super
-          parent.post_add_content
+        def build_control
+          @libui = @parent_proxy.append_quit_item(*@args)
+          handle_listener('on_clicked') do
+            ControlProxy.main_window_proxy&.destroy
+            ::LibUI.quit
+            0
+          end
         end
-        
       end
     end
   end
 end
-
-# TODO Consider moving all shapes underneath Shape namespace
-require 'glimmer/libui/control_proxy/path_proxy'
-require 'glimmer/libui/shape'
-Dir[File.expand_path('../../libui/shape/*.rb', __dir__)].each {|f| require f}