+++ lib/glimmer/libui/control_proxy/text_column_proxy.rb
@@ -19,38 +19,26 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/dsl/expression'
-require 'glimmer/dsl/parent_expression'
+require 'glimmer/libui/control_proxy'
+require 'glimmer/libui/column'
+require 'glimmer/libui/editable_column'
 
 module Glimmer
-  module DSL
-    module Libui
-      class ShapeExpression < Expression
-        include ParentExpression
-  
-        def can_interpret?(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.exists?(keyword) and
-            (
-              parent.is_a?(Glimmer::LibUI::ControlProxy::PathProxy) or
-                parent.is_a?(Glimmer::LibUI::Shape)
-            )
-        end
-  
-        def interpret(parent, keyword, *args, &block)
-          Glimmer::LibUI::Shape.create(keyword, parent, args, &block)
-        end
+  module LibUI
+    class ControlProxy
+      # Proxy for LibUI text column objects
+      #
+      # Follows the Proxy Design Pattern
+      class TextColumnProxy < ControlProxy
+        include Column
+        include EditableColumn
+            
+        private
         
-        def add_content(parent, keyword, *args, &block)
-          super
-          parent.post_add_content
+        def build_control
+          @parent_proxy.append_text_column(name, column_index, editable_value)
         end
-        
       end
     end
   end
 end
-
-# TODO Consider moving all shapes underneath Shape namespace
-require 'glimmer/libui/control_proxy/path_proxy'
-require 'glimmer/libui/shape'
-Dir[File.expand_path('../../libui/shape/*.rb', __dir__)].each {|f| require f}