+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Update default `window` `width` to `190`
 - Improve layout of examples/meta_example.rb
 - Enhance examples/meta_example.rb to enable choosing versions with a spinbox instead of adding them to examples list
+- Nest control proxies under Glimmer::LibUI::ControlProxy namespace
+- Nest shapes under Glimmer::LibUI::Shape namespace
 
 ## 0.1.5